New repository

+++ bookshop-project2/src/main/java/com/feri/alessandro/attsw/project/model/BookDTO.java
@@ -1,13 +1,15 @@
 package com.feri.alessandro.attsw.project.model;
 
+import java.math.BigInteger;
+
 public class BookDTO {
 
-	private Long id;
+	private BigInteger id;
 	private String title;
 	private String author;
 	private double price;
 	
-	public BookDTO(Long id, String title, String author, double price) {
+	public BookDTO(BigInteger id, String title, String author, double price) {
 		super();
 		this.id = id;
 		this.title = title;
@@ -19,29 +21,37 @@ public class BookDTO {
 	
 	}
 
-
-
-	public Long getId() {
+	public BigInteger getId() {
 		return id;
 	}
-	public void setId(Long id) {
+	
+	public void setId(BigInteger id) {
 		this.id = id;
 	}
+	
+	
 	public String getTitle() {
 		return title;
 	}
+	
 	public void setTitle(String title) {
 		this.title = title;
 	}
+	
+	
 	public String getAuthor() {
 		return author;
 	}
+	
 	public void setAuthor(String author) {
 		this.author = author;
 	}
+	
+
 	public double getPrice() {
 		return price;
 	}
+	
 	public void setPrice(double price) {
 		this.price = price;
 	}
